+++ geometry.h
@@ -7,17 +7,17 @@ typedef struct AABB__ AABB;
 #include "communication.h"
 
 struct AABB__ {
-    double min_x;
-    double max_x;
-    double min_y;
-    double max_y;
-    double min_z;
-    double max_z;
+    float min_x;
+    float max_x;
+    float min_y;
+    float max_y;
+    float min_z;
+    float max_z;
 }; //Axis aligned bounding box
 
-double min(double a, double b);
-double max(double a, double b);
-int sgn(double x);
+float min(float a, float b);
+float max(float a, float b);
+int sgn(float x);
 void constructFluidVolume(fluid_particle **fluid_particle_pointers, fluid_particle *fluid_particles, AABB* fluid, int start_x, int number_particles_x, edge *edges, param *params);
 void setParticleNumbers(AABB *boundary_global, AABB *fluid_global, edge *edges, oob *out_of_bounds, int number_particles_x, param *params);
 void partitionProblem(AABB *boundary_global, AABB *fluid_global, int *x_start, int *number_particles_x, param *params);
