+++ glfw_utils.h
@@ -16,7 +16,7 @@ typedef struct {
 static void key_callback(GLFWwindow* window, int key, int scancode, int action, int mods);
 void error_callback(int error, const char* description);
 void check_key_press(GL_STATE_T *state);
-void get_mouse(double *x, double *y, GL_STATE_T *state);
+void get_mouse(float *x, float *y, GL_STATE_T *state);
 void init_ogl(GL_STATE_T *state);
 void exit_ogl(GL_STATE_T *state);
 void swap_ogl(GL_STATE_T *state);
